requirements.txt
argcomplete==2.0.0
asgiref==3.5.0
boto3==1.21.23
botocore==1.24.23
cement==2.8.2
certifi==2021.10.8
cffi==1.15.0
cfn-flip==1.3.0
chardet==4.0.0
charset-normalizer==2.0.12
click==8.0.4
colorama==0.3.9
cryptography==36.0.1
dj-database-url==0.5.0
Django==4.0.2
django-cleanup==6.0.0
django-cors-headers==3.11.0
django-environ==0.8.1
django-extensions==3.1.5
django-rest-knox==4.2.0
django-storages==1.12.3
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
durationpy==0.5
environ==1.0
future==0.16.0
gunicorn==20.1.0
hjson==3.0.2
idna==3.3
jmespath==0.10.0
kappa==0.6.0
lazy-object-proxy==1.4.3
pathspec==0.9.0
Pillow==9.0.1
placebo==0.9.0
pycparser==2.21
PyJWT==2.3.0
pyparsing==2.4.7
python-dateutil==2.8.2
python-slugify==6.1.1
pytz==2021.3
PyYAML==5.4
requests==2.26.0
s3transfer==0.5.2
semantic-version==2.8.5
six==1.11.0
sorl-thumbnail==12.8.0
sqlparse==0.4.2
sslcommerz-lib==1.0
tabulate==0.7.5
termcolor==1.1.0
text-unidecode==1.3
toml==0.10.2
tqdm==4.63.1
troposphere==3.2.2
tzdata==2021.5
urllib3==1.26.8
wcwidth==0.1.9
Werkzeug==2.0.3
whitenoise==6.0.0
wrapt==1.12.1
wsgi-request-logger==0.4.6

